Check if user email and id match before deleting or updating book data

+++ Modules/bookHandler.js
@@ -55,12 +55,27 @@ bookHandler.postBooks = function(request, expressResponse, next){
 bookHandler.updateBooks = function(request, expressResponse, next){
   const {id} = request.params;
   const data = request.body;
+  const userEmail = request.user.email;
+
   // new - returns updated doc instead of old doc
   // overwrite - overwrites doc completely avoiding unwanted properties/side-effects
-  Book.findByIdAndUpdate(id, data, {new: true, overwrite: true})
-    .then(updatedBook => expressResponse.status(200).send(updatedBook))
+  Book.findById(id)
+    .then(book => {
+      if (!book) {
+        return expressResponse.status(404).send('Book not found');
+      }
+
+      if (book.email !== userEmail) {
+        return expressResponse.status(403).send('Unauthorized to update this book');
+      }
+
+      // Proceed with update if the user email matches
+      return Book.findByIdAndUpdate(id, data, { new: true, overwrite: true });
+    })
+    .then(updatedBook => {
+      expressResponse.status(200).send(updatedBook);
+    })
     .catch(err => {
-      // Log error and send a 500 Internal Server Error response
       console.error(err);
       expressResponse.status(500).send('Internal Server Error, updating book');
     });
@@ -69,10 +84,26 @@ bookHandler.updateBooks = function(request, expressResponse, next){
 // Funciton deleteBooks - method of bookHandler
 bookHandler.deleteBooks = function(request, expressResponse, next){
   const {id} = request.params;
-  Book.findByIdAndDelete(id)
-    .then(deletedBook => expressResponse.status(200).send(deletedBook)) // Send back confirmation of deleted data
+  const userEmail = request.user.email;
+
+  // First, find book and check if it belongs to the user by comparing email
+  Book.findById(id)
+    .then(book => {
+      if (!book) {
+        return expressResponse.status(404).send('Book not found');
+      }
+
+      if (book.email !== userEmail) {
+        return expressResponse.status(403).send('Unauthorized to delete this book');
+      }
+
+      // Proceed with deletion if the user email matches
+      return Book.findByIdAndDelete(id);
+    })
+    .then(deletedBook => {
+      expressResponse.status(200).send(deletedBook);
+    })
     .catch(err => {
-      // Log error and send a 500 Internal Server Error response
       console.error(err);
       expressResponse.status(500).send('Internal Server Error, deleting book');
     });
